server/pyproject.toml
[tool.poetry]
name = "resana_server"
version = "0.12.1"
description = ""
authors = ["Scille SAS"]
packages = [
    { include = "antivirus_connector" },
    { include = "asgi_ip_filtering.py" },
    { include = "resana_backend_run.py" },
    { include = "resana_server_entrypoint.py" }
]

[tool.poetry.scripts]
resana_server = 'resana_server_entrypoint:main'

[tool.poetry.dependencies]
python = "~3.9.0"
quart = "^0.18.1"
quart-trio = "^0.10.0"
quart-cors = "^0.5"
hypercorn = "^0.13.2"
httpx = "0.16.0"
oscrypto = "1.3.0"
# install parsec from pypi for linux, or from git for windows & macos
parsec-cloud = [
    { markers = "sys_platform == 'linux'", version = "^2.15.1", extras = ["backend"] },
    { markers = "sys_platform != 'linux'", git = "https://github.com/Scille/parsec-cloud", tag = "v2.15.1" }
]

[tool.poetry.dev-dependencies]
pytest = "^7.0"
black = "^22.6.0"
pytest-trio = "^0.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
